app/requirements.txt
requests
bs4
mysql.connector
fastapi
uvicorn